Update askingQuestions.md
I am not the grammar expert, but I believe "who" is correct in this case. According to Google:

The words "who" and "whom" are both interrogative pronouns, but they are used in different ways. "Who" is used as the subject of a clause, while "whom" is used as the object of a clause. In the sentence "who to ask questions," the pronoun is the subject of the verb "ask," so the correct pronoun to use is "who."

Here are some examples of how to use "who" and "whom" correctly:

Who is the person you are going to ask questions?
To whom should I ask questions about this topic?